% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areas.erDF}
\alias{areas.erDF}
\title{Data table for drawing ER}
\format{A data table with 38,166 rows and 8 variables
\describe{
  \item{long}{longitude of the point}
  \item{lat}{latitude of the point}
  \item{order}{Order by which points are connected}
  \item{hole}{Logical marking if shape drawn is an island or not}
  \item{piece}{id of each unbroken piece}
  \item{id}{ER ID}
  \item{group}{ER ID + id for each unbroken piece}
  ...
}}
\usage{
areas.erDF
}
\description{
A dataset that contains vector information for drawing provinces
}
\keyword{datasets}
