% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areas.provDF}
\alias{areas.provDF}
\title{Data table for drawing provinces}
\format{A data table with 13,205 rows and 8 variables
\describe{
  \item{long}{longitude of the point}
  \item{lat}{latitude of the point}
  \item{order}{Order by which points are connected}
  \item{hole}{Logical marking if shape drawn is an island or not}
  \item{piece}{id of each unbroken piece}
  \item{id}{Province ID}
  \item{group}{Province ID + id for each unbroken piece}
  ...
}}
\usage{
areas.provDF
}
\description{
A dataset that contains vector information for drawing provinces
}
\keyword{datasets}
