% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census.2016.varnames}
\alias{census.2016.varnames}
\title{A named vector for 2016 census column name quick reference}
\format{A named vector with about 200 elements}
\usage{
census.2016.varnames
}
\description{
A named vector for 2016 census column name quick reference
}
\keyword{datasets}
