% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cma.data}
\alias{cma.data}
\title{Meta-data for each CMA/CA}
\format{A data table with 156 rows and 5 variables
\describe{
  \item{ID}{Statistics Canada's code for CMA/CA}
  \item{Name}{English name of the CMA/CA}
  \item{long}{longitude of the centroid points}
  \item{lat}{latitude of the centroid points}
  \item{row.num}{row number for iterating on the colour gradients}
  ...
}}
\usage{
cma.data
}
\description{
A dataset that contains meta data on each of the CMA/CA - used for mapping purposes
}
\keyword{datasets}
