% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{er.data}
\alias{er.data}
\title{Meta-data for each ER}
\format{A data table with 76 rows and 5 variables
\describe{
  \item{ID}{Statistics Canada's code for ER}
  \item{Name}{English name of the ER}
  \item{long}{longitude of the centroid points}
  \item{lat}{latitude of the centroid points}
  \item{row.num}{row number for iterating on the colour gradients}
  ...
}}
\usage{
er.data
}
\description{
A dataset that contains meta data on each of the ER - used for mapping purposes
}
\keyword{datasets}
