% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{export.bf.plot}
\alias{export.bf.plot}
\title{Export a graph into EPS}
\usage{
export.bf.plot(f.name, p.obj, p.height = 6, p.width = 9)
}
\arguments{
\item{f.name}{Character denoting file name, can end with or without .eps Basically a wrapper for ggsave and then embed font}

\item{p.obj}{Plot object to export}

\item{p.height}{height of the plot (in inches)}

\item{p.width}{width of the plot (in inches)}
}
\value{
Exported .EPS with the file name for the given plot object
}
\description{
Export a graph into EPS
}
\examples{
export.bf.plot("Rplot.EPS",plot)
}
