% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.functions.R
\name{plot.change.arrow.bf}
\alias{plot.change.arrow.bf}
\title{Plot a change graph}
\usage{
\method{plot}{change.arrow.bf}(data, x, cat, time.var, order.by = "change",
  annotate = FALSE, label = FALSE, nudge.beg = 0.02, nudge.end = 0.015,
  unit.x = "", x.axis = "", y.axis = "", plot.title = "",
  plot.fig.num = "", caption = "")
}
\arguments{
\item{data}{Data to use}

\item{x}{String. Column name for the X axis - the value the categories plotted are changing}

\item{cat}{String. Column name for the category of things that are changing}

\item{time.var}{String. Column name of variable that denotes time - time has to be numeric}

\item{order.by}{String. One of "change" or "value", way to order the categories. Either by the magnitude of change, or the absolute magnitude}

\item{unit.x}{Character denoting the unit for x-axis. Special formatting for \% and $}

\item{x.axis}{Character denoting the x axis title}

\item{y.axis}{Character denoting the y axis title}

\item{plot.title}{Character denoting title of the plot}

\item{plot.fig.num}{Character denoting plot number (or another plot annotations)}

\item{caption}{Character that denotes caption (Sources)}
}
\value{
ggplot2 object with all the right formatting
}
\description{
Plot a change graph
}
\examples{
plot.change.arrow.bf(change.gdp,"gdp","country","year")
}
