% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.functions.R
\name{plot.column.bf}
\alias{plot.column.bf}
\title{Create a simple bar chart with Brookfield theme. Uses Rooney font family - make sure you install that with extrafonts before using this.}
\usage{
plot.column.bf(data,x,cat,plot.title="",plot.fig.num="",
order.bar = FALSE, group.by = "", label = FALSE, export = FALSE,
export.name = "")
}
\arguments{
\item{data}{Main data table to plot}

\item{x}{Character corresponding the column name for what's being counted in the bar graph}

\item{cat}{Character corresponding to the column name categories of the column}

\item{order.bar}{One of "none", "ascending", or "descending"}

\item{group.by}{Character corresponding to the column name that columns are grouped by (colour)}

\item{colours}{Vector (or set.colour function) of colours to use. If not, default palette is generated.}

\item{stacked}{TRUE/FALSE to stack the columns or not still supported but maybe not in the future}

\item{label}{TRUE/FALSE on whether to label the points or not}

\item{label.unit}{What unit does y axis (and labels if any) will have}

\item{label.adjust}{Factor to adjust the labels by default is 2.5\% or 0.025}

\item{plot.title}{Character denoting title of the plot}

\item{plot.fig.num}{Character denoting plot number (or another plot annotations)}

\item{y.axis}{Character denoting axis title for y (unit)}

\item{legend.title}{Title for the legend if multiple colours are used}

\item{caption}{Character denoting sources and other captions}

\item{export}{TRUE or FALSE whether to export file as EPS under default options (height=6 inches, width=9 inches)}

\item{export.name}{Name of the exported EPS file}

\item{position}{One of "identity", "stacked" or "dodge"}
}
\value{
Single column plot that conforms to Brookfield styling
}
\description{
Create a simple bar chart with Brookfield theme. Uses Rooney font family - make sure you install that with extrafonts before using this.
}
\examples{
plot.column.bf(data,"income","automation.risk")
}
