% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.functions.R
\name{plot.line.bf}
\alias{plot.line.bf}
\title{Plot a line chart
Make sure you don't confuse between x and y. X is what goes on the bottom (Year...)}
\usage{
\method{plot}{line.bf}(data, x, y, group.by = NULL, show.points = FALSE,
  cat.order = NULL, ingraph.labels = FALSE, colours = NULL,
  plot.title = "", plot.fig.num = "", plot.limit = NULL, unit.x = "",
  unit.y = "", x.axis = "", y.axis = "", legend.title = "",
  caption = "", export = FALSE, export.name = "")
}
\arguments{
\item{data}{Data to use}

\item{x}{The X axis the line will be connected along this point}

\item{y}{The Y axis points will reside along this}

\item{group.by}{Line colour will be determined by this}

\item{show.points}{TRUE/FALSE whether to show the points on the line or not}

\item{cat.order}{Put the order of categorical variable in here (as a vector) c("cat.1","cat.2","cat.3") and so on.
If you don't, a default order will be generated.}

\item{ingraph.labels}{TRUE/FALSE whether to show a label at the end of the line}

\item{colours}{Vector (or set.colour function) of colours to use. If not, default palette is generated.}

\item{plot.title}{Character denoting title of the plot}

\item{plot.fig.num}{Character denoting plot number (or another plot annotations)}

\item{plot.limit}{Vector of form c(x.min,x.max,y.min,y.max)}

\item{unit.x}{Character denoting the unit for x-axis. Special formatting for \% and $}

\item{unit.y}{Character denoting the unit for y-axis. Special formatting for \% and $}

\item{x.axis}{Character denoting the x axis title}

\item{y.axis}{Character denoting the y axis title}

\item{legend.title}{Character denoting the title for the colour legend}

\item{caption}{Character that denotes caption (Sources)}

\item{export}{TRUE/FALSE whether to export file as EPS under default options
(height=scaled according to row number and cell number, width=12 inches)}

\item{export.name}{Name of the exported EPS file}
}
\value{
ggplot2 object with all the right formatting
}
\description{
Plot a line chart
Make sure you don't confuse between x and y. X is what goes on the bottom (Year...)
}
\examples{
plot.line.bf(data,"Year","Income",group.by="Gender")
plot.line.bf(data,"Education","Count",cat.order = c("High School","Bachelor's","Master's","PhD"))
}
