% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.functions.R
\name{plot.map.csd.bf}
\alias{plot.map.csd.bf}
\title{Create a map of selected Canadian CSD with shades}
\usage{
plot.map.csd.bf(province.name = c("NL","PE","NS","NB","QC","ON","MB","SK","AB","BC","YT","NT","NU"),
value.data = NULL,
colours = NULL,
plot.title = "",
plot.fig.num = "",
caption = "",
legend.title = "")
}
\arguments{
\item{province.name}{A vector of the 2 letter code for provinces to plot. One or more of:
"NL","PE","NS","NB","QC","ON","MB","SK","AB","BC","YT","NT","NU"}

\item{value.data}{a data table with 2 columns - first column must be string that denotes Province-CSD ID. Second column has to be numeric and is the value you want to show}

\item{colours}{Vector (or set.colour function) of colours to use. If not, default palette is generated.}

\item{plot.title}{Character denoting title of the plot}

\item{plot.fig.num}{Character denoting plot number (or another plot annotations)}

\item{caption}{character for caption (sources etc)}

\item{legend.title}{Character denoting legend titles}

\item{export}{TRUE/FALSE whether to export file as EPS under default options (height=6 inches, width=9 inches)}

\item{export.name}{Name of the exported EPS file}
}
\value{
A map plot that conforms to Brookfield style
}
\description{
Create a map of selected Canadian CSD with shades
}
\examples{
plot.map.csd.bf()
}
