% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.functions.R
\name{plot.mekko.bf}
\alias{plot.mekko.bf}
\title{Plot a mekko chart}
\usage{
\method{plot}{mekko.bf}(data, x, y, plot.title = "", plot.fig.num = "",
  caption = "", group.by = NULL, legend.title = "", colours = NULL,
  unit.x = "", unit.y = "", x.axis = "", y.axis = "", export = FALSE,
  export.name = "")
}
\arguments{
\item{data}{Data to use}

\item{x}{String. The X axis - the values that the width of each bar will represent}

\item{y}{String. The Y axis - what the height of bars will represent}

\item{plot.title}{Character denoting title of the plot}

\item{plot.fig.num}{Character denoting plot number (or another plot annotations)}

\item{caption}{Character that denotes caption (Sources)}

\item{group.by}{String. Variable that the bar should be coloured by}

\item{legend.title}{Character denoting the title for the colour legend}

\item{colours}{Vector (or set.colour function) of colours to use. If not, default palette is generated.}

\item{unit.x}{Character denoting the unit for x-axis. Special formatting for \% and $}

\item{unit.y}{Character denoting the unit for y-axis. Special formatting for \% and $}

\item{x.axis}{Character denoting the x axis title}

\item{y.axis}{Character denoting the y axis title}

\item{export}{TRUE/FALSE whether to export file as EPS under default options
(height=scaled according to row number and cell number, width=12 inches)}

\item{export.name}{Name of the exported EPS file}
}
\value{
ggplot2 object with all the right formatting
}
\description{
Plot a mekko chart
}
\examples{
plot.mekko.bf(mtcars,"mpg","disp")
}
