% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.functions.R
\name{plot.pyramid.bf}
\alias{plot.pyramid.bf}
\title{Plot a pyramid
Make sure diff is 2 groups and x has the aggregated value}
\usage{
\method{plot}{pyramid.bf}(data, x, diff, group, group.order = NULL,
  colours = NULL, plot.title = "", plot.fig.num = "", unit.x = "",
  x.axis = "", y.axis = "", legend.title = "", caption = "",
  export = FALSE, export.name = "")
}
\arguments{
\item{data}{Data to use}

\item{x}{The X axis - the values that the bars will represent}

\item{diff}{The 2 groups the bar will contrast}

\item{group}{All the different groups that the ar will show (age groups)}

\item{group.order}{Order in with the group is displayed in}

\item{colours}{Vector (or set.colour function) of colours to use. If not, default palette is generated.}

\item{plot.title}{Character denoting title of the plot}

\item{plot.fig.num}{Character denoting plot number (or another plot annotations)}

\item{unit.x}{Character denoting the unit for x-axis. Special formatting for \% and $}

\item{x.axis}{Character denoting the x axis title}

\item{y.axis}{Character denoting the y axis title}

\item{legend.title}{Character denoting the title for the colour legend}

\item{caption}{Character that denotes caption (Sources)}

\item{export}{TRUE/FALSE whether to export file as EPS under default options
(height=scaled according to row number and cell number, width=12 inches)}

\item{export.name}{Name of the exported EPS file}
}
\value{
ggplot2 object with all the right formatting
}
\description{
Plot a pyramid
Make sure diff is 2 groups and x has the aggregated value
}
\examples{
plot.pyramid.bf(data,"Count","Gender","Age")
}
