% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.functions.R
\name{plot.scatter.bf}
\alias{plot.scatter.bf}
\title{Create a scatter plot with Brookfield formatting. Uses Rooney font family - make sure you install that with extrafonts before using this.}
\usage{
plot.scatter.bf(data,x,y,group.by=NULL,p.size=NULL,unit.x="",
unit.y="",plot.title="",plot.fig.num="",title.x.axis="",
title.y.axis="",legend.title = "",deg.45 = FALSE,
export = FALSE,export.name = "")
}
\arguments{
\item{data}{Main data table to plot}

\item{x}{Character corresponding the column name for x-axis}

\item{y}{Character corresponding to the column name for y-axis}

\item{group.by}{Character corresponding to the column name that points are grouped by (colour)}

\item{colours}{Vector (or set.colour function) of colours to use. If not, default palette is generated.}

\item{p.size}{Character corresponding to the column name that sets the size for points}

\item{deg.45}{TRUE/FALSE denoting whether to add 45 degree line or not}

\item{trend.line}{TRUE/FALSE add a trend line using OLS}

\item{plot.limit}{Vector of form c(min.x,max.x,min.y,max.y) that sets out plot limits}

\item{unit.x}{Character denoting the unit for x-axis. Special formatting for \% and $}

\item{unit.y}{Character denoting the unit for y-axis. Special formatting for \% and $}

\item{x.axis}{Character denoting title for the x axis}

\item{y.axis}{Character denoting title for the y axis}

\item{plot.title}{Character denoting title of the plot}

\item{plot.fig.num}{Character denoting plot number (or another plot annotations)}

\item{caption}{character for caption (sources etc)}

\item{legend.title}{Character denoting legend titles}

\item{export}{TRUE/FALSE whether to export file as EPS under default options (height=6 inches, width=9 inches)}

\item{export.name}{Name of the exported EPS file}
}
\value{
Scatter plot that conforms to Brookfield style
}
\description{
Create a scatter plot with Brookfield formatting. Uses Rooney font family - make sure you install that with extrafonts before using this.
}
\examples{
plot.scatter.bf(data,"income","automation.risk")
}
