% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.functions.R
\name{plot.waffle.bf}
\alias{plot.waffle.bf}
\title{Plot a waffle chart with dividing lines and labels
This function requires you to know the distribution of the squares already.
In the future there will be a wrapper that helps you with determining the proportions}
\usage{
\method{plot}{waffle.bf}(named.vector, row.num = 5, colours = NULL,
  plot.title = "", plot.fig.num = "", plot.cat.title = "", caption = "",
  x.axis = "", dividing.line = FALSE, labels = FALSE, label.unit = "",
  export = FALSE, export.name = "Rplot")
}
\arguments{
\item{named.vector}{A named vector that is of the form c("label"=num.square)}

\item{row.num}{Number of rows to put in a waffle chart}

\item{colours}{Vector (or set.colour function) of colours to use. If not, default palette is generated.}

\item{plot.title}{Character denoting title of the plot}

\item{plot.fig.num}{Character denoting plot number (or another plot annotations)}

\item{plot.cat.title}{Legend title for the colour of the plot}

\item{caption}{String that sets the caption: Sources, notes stc}

\item{x.axis}{Notes to be placed - Recommended place to put what each square means if not clear already}

\item{dividing.line}{TRUE/FALSE on whether we want to display the dividing lines between the groups}

\item{labels}{TRUE/FALSE on whether we want to have labels on them or not}

\item{label.unit}{character string that can add labels to it. Examples include "\%"}

\item{export}{TRUE/FALSE whether to export file as EPS under default options
(height=scaled according to row number and cell number, width=12 inches)}

\item{export.name}{Name of the exported EPS file}
}
\value{
ggplot2 object with all the right formatting
}
\description{
Plot a waffle chart with dividing lines and labels
This function requires you to know the distribution of the squares already.
In the future there will be a wrapper that helps you with determining the proportions
}
\examples{
plot.waffle.bf(c("Red"=20,"Blue"=30,"Green"=50),label.unit="\%")
}
