% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province.data}
\alias{province.data}
\title{Meta-data for each provinces}
\format{A data table with 13 rows and 6 variables
\describe{
  \item{code}{Statistics Canada's code for provinces}
  \item{province}{English name of the provinces}
  \item{region}{Statistics Canada's region for the province}
  \item{abbrev}{2 letter abbreviation for the provinces}
  \item{long}{longitude of the centroid points}
  \item{lat}{latitude of the centroid points}
  \item{row.num}{row number for iterating on the colour gradients}
  ...
}}
\source{
\url{https://www.statcan.gc.ca/pub/92-195-x/2011001/geo/prov/tbl/tbl8-eng.htm}
}
\usage{
province.data
}
\description{
A dataset that contains meta data on each of the provinces - used for mapping purposes
}
\keyword{datasets}
