% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{set.colours}
\alias{set.colours}
\title{Set a colour palette}
\usage{
set.colours(n)
}
\arguments{
\item{n}{Integer to select the right set of colours}
}
\value{
Character vector containing HEX value for colours
}
\description{
Set a colour palette
}
\examples{
set.colours(1)
}
