% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{set.ticks.seq}
\alias{set.ticks.seq}
\title{Set appropriate ticks. Wrapper for cbreaks in scale that allows for $ and percent ticks}
\usage{
set.ticks.seq(max, min, unit, num.ticks = 5)
}
\arguments{
\item{max}{numerical max of the ticked plot}

\item{min}{numerical min of the ticked plot}

\item{unit}{Unit that goes wit hthe ticks}
}
\value{
A vector with $breaks which are breaks and $labels which are the labelled values
}
\description{
Set appropriate ticks. Wrapper for cbreaks in scale that allows for $ and percent ticks
}
\examples{
set.ticks.seq(100,0,"\%")
}
