% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{export.bf.plot}
\alias{export.bf.plot}
\title{Export a graph into PDF}
\usage{
export.bf.plot(f.name, p.obj, p.height = 6, p.width = 7.25)
}
\arguments{
\item{f.name}{Character denoting file name, can end with or without .pdf Basically a wrapper for ggsave and then embed font}

\item{p.obj}{Plot object to export}

\item{p.height}{height of the plot (in inches)}

\item{p.width}{width of the plot (in inches)}
}
\value{
Exported .PDF with the file name for the given plot object
}
\description{
Export a graph into PDF
}
\examples{
export.bf.plot("Rplot.PDF",plot)
}
