% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{set.colours}
\alias{set.colours}
\title{Set a colour palette or generate one according to your specification.
Refers to the main Style Guide on choosing colours that match your intention and report style.
set.colours <- function(n,type = "categorical",gradient.choice = "dark.blue")}
\usage{
set.colours(n, type = "categorical", gradient.choice = "dark.blue",
  categorical.choice = NULL, special = NULL)
}
\arguments{
\item{n}{Integer to select the right set of colours}

\item{type}{Either "categorical" or "gradient", decides the type of colour you're going to get}

\item{gradient.choice}{For gradient type, specify the colour. One of: c("dark.blue","light.blue",
"pink","yellow","magenta","orange","green","teal","grey")}

\item{categorical.choice}{Specify custom colour names as a vector (same name as gradient.choice)One of:
c("dark.blue","light.blue","pink","yellow","magenta","orange","green","teal","grey")}

\item{special}{special colours, will be added in the future - opposing colours, opposing gradients}
}
\value{
Character vector containing HEX value for colours
}
\description{
Set a colour palette or generate one according to your specification.
Refers to the main Style Guide on choosing colours that match your intention and report style.
set.colours <- function(n,type = "categorical",gradient.choice = "dark.blue")
}
\examples{
set.colours(1,type="categorical",gradient.choice="dark.blue")
}
